test: extract HostFS white-box suite
